armada-8k: vfio: xhci: add support for XHCI reset hook
The VFIO reset hook is called every time a platform device is
passed to a guest or removed from a guest.
When the XHCI device is unbound from the host, the host driver
disables the XHCI phys so when the device is passed to the
guest it becomes dis-functional.
This patch uses the VFIO reset hook to enable
the XHCI phys on behalf of the guest.

Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Signed-off-by: Tomasz Nowicki <tn@semihalf.com>
Change-Id: I932a76001a1cdb487d13d5ae749e3ad373eccba5