vfio: xhci: update support for XHCI reset hook
The VFIO reset hook is called every time a platform device is
passed to a guest or removed from a guest.
When the XHCI device is unbound from the host, the host driver
disables the XHCI phys so when the device is passed to the
guest it becomes dis-functional.
This patch uses the VFIO reset hook to enable
the XHCI phys on behalf of the guest.

Change-Id: I59de5de33bc59542f70db6bdc2f88728cfe8df0c
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/56407
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>