iommu/io-pgtable-arm-v7s: Abort allocation when table address overflows the PTE
When run on a 64-bit system, the v7s driver may obtain page table with
physical addresses larger than 32-bit. Currently map() truncates the
address and writes it into a PTE, leading to iova_to_phys or unmap
accessing invalid memory. They may access MMIO space, which simply fails
the test in my model but could as well crash the CPU or have subtle side
effects. Kasan reports it as a use-after-free. Test if the physical
address fits in a PTE before returning a new table.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>
---

I'm not sure how to fix this properly, or if it even deserves a fix
(doubt anyone in production enables v7s selftest on 64-bit systems,
which might be the only way to trigger this), hence the RFC tag. Is
there a way to enforce 32-bit allocation?