virtio/iommu: Add IDENTITY regions
This is work in progress, but I think it's right way. Host sets up a
region for SW MSIs at 0x800000. Comunicates an identity region to guest.

           VA       S1      GPA      S2       PA
	0x8000000 -----> 0x8000000 -----> GIC doorbell

But guest has a virtual GIC, and needs the guest to use that for setting
up MSIs. The host should report it as HW MSI region! Otherwise the guest
will put an IOVA in the MSI-X table, and the host, which doesn't know how
to translate this, doesn't know which MSI doorbell it is. Also given that
it's a static region in Linux the VA will be the same in host and guest,
leading to a conflict.

If the host-guest interface is a HW region, there still is a risk that
GPA(vgic) == GPA(gic). It's really the host's problem, because it
obviously shouldn't put devices in the reserved regions of the GPA map,
that are reported by the host kernel.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>