KVM: arm64: nv: Try to gracefully handle HCR_EL2.E2H flips
The ARMv8.4-NV extension, for all its benefits, has a few flaws. One of
them is the fact that we can't detech changes to HCR_EL2.E2H, given that
accesses to HCR_EL2 do not trap.

This means that if the guest flips things behind our back, we may end-up
with the wrong state, specially if the guest is turning E2H off.

In order to deal with this, let's pray that we'll have a trapping
instruction right after the write to HCR_EL2, and detect that it has
changed. If so, undo the write(!), and postpone it until we've fully
exited by using a new request. We then perform the write as if we were
on ARMv8.3-NV, which implies the put/load dance.

This guarantees that on rentry of the guest, we'll have the right trap
flags. That is, if we trapped the first place... :-/ For this reason,
this should at best be considered a very sorry hack.

Signed-off-by: Marc Zyngier <maz@kernel.org>