KVM: arm64: Only switch MDSCR_EL1 when KVM_ARM64_DEBUG_DIRTY is set
We currentlt eagerly switch MDSCR_EL1 at all times, irrespective of the
state of KVM_ARM64_DEBUG_DIRTY. This is a bit odd, as most of the time,
this register is trapped, and thus cannot change behind our back.

It is only when KVM_ARM64_DEBUG_DIRTY is set that context-switching
MDSCR_EL1 becomes important. So let's move this save/restore operation
to the debug code, which only gets triggered when the flag is set.

On top of reducing the work we do on the critical path, it also reduces
the number of traps that are generated by a nested KVM if the host
hypervisor has MDCR_EL2.{TDA,TDE} != {0,0}.

Signed-off-by: Marc Zyngier <maz@kernel.org>