KVM: arm/arm64: Fix spurious soft timer kick for emulated timer
The emulated timer needs no further software timer if the timer can fire
now and no change in line level: (should_fire == 1 &&
should_fire == ctx->irq.level). In that case htimer should be simply
canceled.

This was found with NV testing where we have at most 4 timers and the two
are purely emulated resulting in the guest getting stuck in an interrupt
storm.

Fixes: bee038a674875 ("KVM: arm/arm64: Rework the timer code to use a timer_map")
Signed-off-by: Tomasz Nowicki <tnowicki@marvell.com>