vfio: platform: Support VFIO platform clock handling
The main idea behind the VFIO clock extension is to use emulated clock
for guest and trap to hypevisor whenever host supposed to change real
clock configuration. As guest will trap to userspace emulation tool
(QEMU most likely) we need API extension so that we can initialize,
query and configure clock like we already do for interrupts and memory
regions.

The new VFIO interace allow to:
- initialize device clocks (clk_get() for all possible clocks) when
  VFIO platform device file is opened
- get number of possible clock with VFIO_DEVICE_GET_INFO ioctl
- manage clock by calling host clock interface
  (clk_{enable|disable|prepare|unprepare|set_rate|get_rate}) with
  VFIO_DEVICE_CLK ioctl and respective flags

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>