vfio: platform: Initialize VFIO platform clock handling
The main idea behind the VFIO clock extension is to use emulated clock
in guest and handle guest requests whenever host supposed to change real
clock configuration. We need API extension so that we can initialize,
query and configure clock like we already do for interrupts and memory
regions.

The first step of VFIO clock handling is to:
- initialize device clocks (clk_get() for all possible clocks) when
  VFIO platform device file is opened
- get number of possible clock with VFIO_DEVICE_GET_INFO ioctl

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>
Change-Id: Ibdd16c08d64bd25423e941e30ba4914c722a7608